Move all display functions to a class
This is important because it allowed me to pull the display and it's
canvases out of the global scope. I've reverted to using the old school
"new" keyword to allocate objects on the heap. I'm doing this in
preperation of trying to figure out how to move them to the SPRAM that
the FeatureS2 offer.